Only active documentation units
Procedure list only contains the active documentation units (documentation unit with the procedure as last set entry).

RISDEV-2857

+++ backend/src/main/java/de/bund/digitalservice/ris/caselaw/domain/Procedure.java
@@ -5,4 +5,4 @@ import java.util.UUID;
 import lombok.Builder;
 
 @Builder
-public record Procedure(UUID id, String label, Integer documentUnitCount, Instant createdAt) {}
+public record Procedure(UUID id, String label, Long documentUnitCount, Instant createdAt) {}
